REFACTOR: put two origin evaluators into one file and add RankEvaluator
