+++ recbole/evaluator/__init__.py
@@ -1,4 +1,4 @@
 from recbole.evaluator.abstract_evaluator import *
-from recbole.evaluator.loss_evaluator import *
+from recbole.evaluator.proxy_evaluator import *
 from recbole.evaluator.metrics import *
-from recbole.evaluator.topk_evaluator import *
+from recbole.evaluator.evaluators import *
